+++ kernel/pf_q-bits.h
@@ -26,9 +26,9 @@
  
 
 #define pfq_ctz(n) \
-	 __builtin_choose_expr(__builtin_types_compatible_p(typeof(n),unsigned int), __builtin_ctz(n), \
-         __builtin_choose_expr(__builtin_types_compatible_p(typeof(n),unsigned long), __builtin_ctzl(n), \
-         __builtin_choose_expr(__builtin_types_compatible_p(typeof(n),unsigned long long), __builtin_ctzll(n), (void)0 )))
+	 __builtin_choose_expr(__builtin_types_compatible_p(typeof(n),unsigned int),            (unsigned int)__builtin_ctz(n), \
+         __builtin_choose_expr(__builtin_types_compatible_p(typeof(n),unsigned long),       (unsigned int)__builtin_ctzl(n), \
+         __builtin_choose_expr(__builtin_types_compatible_p(typeof(n),unsigned long long),  (unsigned int)__builtin_ctzll(n), (void)0 )))
 
 #define bitwise_for_each(mask, n) \
 	for(; n = pfq_ctz(mask), mask ; mask^=(1UL << n))
